+++ deepl-java/src/main/java/com/deepl/api/Translator.java
@@ -82,7 +82,7 @@ public class Translator {
    */
   private String constructUserAgentString(boolean sendPlatformInfo, AppInfo appInfo) {
     StringBuilder sb = new StringBuilder();
-    sb.append("deepl-java/1.6.0");
+    sb.append("deepl-java/1.7.0");
     if (sendPlatformInfo) {
       sb.append(" (");
       Properties props = System.getProperties();