+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [1.7.0] - 2024-11-15
 ### Added
 * Added `modelType` option to `translateText()` to use models with higher
   translation quality (available for some language pairs), or better latency.
@@ -125,6 +125,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
+[1.7.0]: https://github.com/DeepLcom/deepl-java/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/DeepLcom/deepl-java/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/DeepLcom/deepl-java/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/DeepLcom/deepl-java/compare/v1.4.0...v1.5.0